+++ src/Control/Effect/FossaApiClient.hs
@@ -46,6 +46,7 @@ module Control.Effect.FossaApiClient (
   deleteReleaseGroup,
   deleteReleaseGroupRelease,
   createReleaseGroup,
+  createReleaseGroupRelease,
   getReleaseGroups,
   getReleaseGroupReleases,
   updateReleaseGroupRelease,
@@ -59,7 +60,7 @@ import App.Fossa.VSI.Fingerprint qualified as Fingerprint
 import App.Fossa.VSI.IAT.Types qualified as IAT
 import App.Fossa.VSI.Types qualified as VSI
 import App.Fossa.VendoredDependency (VendoredDependency)
-import App.Types (FullFileUploads, ProjectMetadata, ProjectRevision)
+import App.Types (FullFileUploads, ProjectMetadata, ProjectRevision, ReleaseGroupReleaseRevision)
 import Container.Types qualified as NativeContainer
 import Control.Algebra (Has)
 import Control.Carrier.Simple (Simple, sendSimple)
@@ -143,6 +144,7 @@ data FossaApiClientF a where
   DeleteReleaseGroup :: Int -> FossaApiClientF ()
   DeleteReleaseGroupRelease :: Int -> Int -> FossaApiClientF ()
   CreateReleaseGroup :: CreateReleaseGroupRequest -> FossaApiClientF CreateReleaseGroupResponse
+  CreateReleaseGroupRelease :: Int -> ReleaseGroupReleaseRevision -> FossaApiClientF ReleaseGroupRelease
   UpdateReleaseGroupRelease :: Int -> Int -> UpdateReleaseRequest -> FossaApiClientF ReleaseGroupRelease
   GetReleaseGroups :: FossaApiClientF [ReleaseGroup]
   GetReleaseGroupReleases :: Int -> FossaApiClientF [ReleaseGroupRelease]
@@ -289,6 +291,9 @@ updateReleaseGroupRelease releaseGroupId releaseId updateReq = sendSimple $ Upda
 
 createReleaseGroup :: Has FossaApiClient sig m => CreateReleaseGroupRequest -> m CreateReleaseGroupResponse
 createReleaseGroup req = sendSimple $ CreateReleaseGroup req
+
+createReleaseGroupRelease :: Has FossaApiClient sig m => Int -> ReleaseGroupReleaseRevision -> m ReleaseGroupRelease
+createReleaseGroupRelease releaseGroupId req = sendSimple $ CreateReleaseGroupRelease releaseGroupId req
 
 getReleaseGroups :: Has FossaApiClient sig m => m [ReleaseGroup]
 getReleaseGroups = sendSimple GetReleaseGroups